AutoPilot-CSV-Gen.ps1
AutoPilot-CSV-Gen.ps1

This file aims to grab the hardware hash from a computer and then create the csv file that is utilized for adding a computer into Microsoft Intune AutoPilot. In essence you run this script, get the CSV, log into Intune and Import the CSV and the device is imported into AutoPilot.